📦 refactor: move test-only formatters and types out of application code, reject invalid usernames
- Moved formatUsername and isoStringFieldsToDate from src/utils into src/tests/_utils — both are only ever used by test fixtures/assertions, never by application code, so src/utils no longer has a reason to exist and was removed entirely
- Moved TDateToString, TStringToDate and TApiResponse from src/types/generics.ts into src/tests/_types/generics.ts for the same reason; TDataInput and TGetDto stay in src/types since they're used by repositories and dtos
- createUserSchema now rejects any username with characters other than letters/digits via a regex, instead of silently transforming input into something the user didn't type (formatUsername stays a test-fixture helper for turning faker's free-form usernames into valid ones)
- Removed the standalone unit test for formatUsername, since it's a test utility and doesn't need its own coverage
