+++ src/models/user-model/schema.ts
@@ -1,7 +1,12 @@
 import { z } from 'zod'
 
 export const createUserSchema = z.object({
-  username: z.string({ message: 'O username é obrigatório.' }),
+  username: z
+    .string({ message: 'O username é obrigatório.' })
+    .min(1, { message: 'O username é obrigatório.' })
+    .regex(/^[a-zA-Z0-9]+$/, {
+      message: 'O username deve conter apenas letras e números.'
+    }),
   email: z.email({ message: 'O email informado não é válido.' }),
   password: z
     .string({ message: 'A senha é obrigatória.' })
