+++ src/tests/integration/_use-cases/registration-flow.test.ts
@@ -3,9 +3,9 @@ import setCookieParser from 'set-cookie-parser'
 import { API_BASE_URL, APP_BASE_URL } from '@/constants'
 import { TSessionDto, TUserDto } from '@/dtos'
 import { activationModel, userModel } from '@/models'
+import { TApiResponse } from '@/tests/_types'
+import { isoStringFieldsToDate } from '@/tests/_utils'
 import orchestrator from '@/tests/orchestrator'
-import { TApiResponse } from '@/types'
-import { isoStringFieldsToDate } from '@/utils'
 
 beforeAll(async () => {
   await orchestrator.waitForAllServices()
