+++ src/tests/integration/api/v1/activations/[token_id]/patch.test.ts
@@ -9,9 +9,9 @@ import {
   userModel
 } from '@/models'
 import { TUserActivationToken } from '@/repositories'
+import { TApiResponse } from '@/tests/_types'
+import { isoStringFieldsToDate } from '@/tests/_utils'
 import orchestrator from '@/tests/orchestrator'
-import { TApiResponse } from '@/types'
-import { isoStringFieldsToDate } from '@/utils'
 
 beforeAll(async () => {
   await orchestrator.waitForAllServices()
