+++ src/tests/integration/api/v1/sessions/post.test.ts
@@ -5,9 +5,9 @@ import { API_BASE_URL } from '@/constants'
 import { TSessionDto } from '@/dtos'
 import { TErrorResponse } from '@/infra'
 import { SESSION_TOKEN_EXPIRATION_IN_MILLISECONDS } from '@/models'
+import { TApiResponse } from '@/tests/_types'
+import { isoStringFieldsToDate } from '@/tests/_utils'
 import orchestrator from '@/tests/orchestrator'
-import { TApiResponse } from '@/types'
-import { isoStringFieldsToDate } from '@/utils'
 
 beforeAll(async () => {
   await orchestrator.waitForAllServices()
