+++ src/tests/integration/api/v1/status/get.test.ts
@@ -1,8 +1,8 @@
 import { API_BASE_URL } from '@/constants'
 import { TStatusDto } from '@/dtos'
+import { TApiResponse } from '@/tests/_types'
+import { isoStringFieldsToDate } from '@/tests/_utils'
 import orchestrator from '@/tests/orchestrator'
-import { TApiResponse } from '@/types'
-import { isoStringFieldsToDate } from '@/utils'
 
 beforeAll(async () => {
   await orchestrator.waitForAllServices()
