+++ src/tests/integration/api/v1/user/get.test.ts
@@ -8,9 +8,9 @@ import {
   SESSION_TOKEN_EXPIRATION_IN_MILLISECONDS,
   sessionModel
 } from '@/models'
+import { TApiResponse } from '@/tests/_types'
+import { isoStringFieldsToDate } from '@/tests/_utils'
 import orchestrator from '@/tests/orchestrator'
-import { TApiResponse } from '@/types'
-import { isoStringFieldsToDate } from '@/utils'
 
 beforeAll(async () => {
   await orchestrator.waitForAllServices()
