+++ src/tests/integration/api/v1/users/[username]/patch.test.ts
@@ -5,9 +5,9 @@ import { TUserDto } from '@/dtos'
 import { TErrorResponse } from '@/infra'
 import { passwordModel } from '@/models'
 import { userRepository } from '@/repositories'
+import { TApiResponse } from '@/tests/_types'
+import { isoStringFieldsToDate } from '@/tests/_utils'
 import orchestrator from '@/tests/orchestrator'
-import { TApiResponse } from '@/types'
-import { isoStringFieldsToDate } from '@/utils'
 
 beforeAll(async () => {
   await orchestrator.waitForAllServices()
