+++ src/tests/integration/api/v1/users/post.test.ts
@@ -4,9 +4,9 @@ import { API_BASE_URL } from '@/constants'
 import { TUserDto } from '@/dtos'
 import { TErrorResponse } from '@/infra'
 import { passwordModel, userModel } from '@/models'
+import { TApiResponse } from '@/tests/_types'
+import { isoStringFieldsToDate } from '@/tests/_utils'
 import orchestrator from '@/tests/orchestrator'
-import { TApiResponse } from '@/types'
-import { isoStringFieldsToDate } from '@/utils'
 
 beforeAll(async () => {
   await orchestrator.waitForAllServices()
